Split server.js into two files to allow for testing with supertest npm package
