+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 coverage/
 server/public/bundle.js
 .env
+.test.env
